8 create filter (#20)
* use long as id in reposiotries

* modify creation date in test data

* use long as id in repository

* add filter logic in service

* add filter logic in service

+++ src/main/java/my/photomanager/photo/category/PhotoCategoryRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import lombok.NonNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhotoCategoryRepository extends JpaRepository<PhotoCategory, Integer> {
+public interface PhotoCategoryRepository extends JpaRepository<PhotoCategory, Long> {
 
 	Optional<PhotoCategory> findByName(@NonNull String name);
 }
