+++ src/main/java/my/photomanager/photo/location/PhotoLocationRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import lombok.NonNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhotoLocationRepository extends JpaRepository<PhotoLocation, Integer> {
+public interface PhotoLocationRepository extends JpaRepository<PhotoLocation, Long> {
 
 	Optional<PhotoLocation> findByCountryAndCity(@NonNull String country, @NonNull String city);
